requirements.txt
# openai>=1.0.0
# flask
# streamlit

flask
google-generativeai
